Update Factory Machines.cpp

+++ CSES/Factory Machines.cpp
@@ -1,8 +1,8 @@
 #include <bits/stdc++.h>
 using namespace std;
 
-int bin_search(int n, int t, vector<int>& k){
-    int l = 0, r = t*k[0], mid = (l+r)/2, mm = INT_MAX;
+long long bin_search(long long n, long long t, vector<long long>& k){
+    long long l = 0, r = t*k[0], mid = (l+r)/2, mm = LLONG_MAX;
 
     while(l < r){
         mid = (l+r)/2;
@@ -20,9 +20,9 @@ int bin_search(int n, int t, vector<int>& k){
 }
 
 int main(){
-    int n,t;
+    long long n,t;
     cin >> n >> t;
-    vector<int> k(n);
+    vector<long long> k(n);
 
     for(int i = 0; i < n; i++) cin >> k[i];
     sort(k.begin(), k.end());
